Add missing foreign keys in db

+++ app/models/review/Review.scala
@@ -2,7 +2,7 @@ package models.review
 
 import play.api.libs.json.Json
 
-case class Review(id: Long, grade: Int, comment: String, productId: Int)
+case class Review(id: Long, grade: Int, comment: String, productId: Long)
 
 object Review {
   implicit val reviewJSON = Json.format[Review]
